verify extra parts
